requirements.txt
# local package
-e .

# external packages
tensorflow-gpu==2.1.0
tensorboard==2.1.0
h5py==2.10.0
biopython==1.76
ipykernel
joblib==0.14.1
matplotlib==3.1.3
numpy==1.18.1
pandas==1.0.1
Pillow==7.0.0
pyteomics==4.2
pytest==5.3.5
scikit-learn==0.22.1
scipy==1.5.0
statsmodels==0.11.0
tqdm==4.42.1
requests==2.22.0
requests-futures==1.0.0
seaborn==0.10.0
pip==20.0.2

# external requirements (cookiecutter data science)
# click
# Sphinx
# coverage
# awscli
# flake8
# python-dotenv>=0.5.1
